migrate and fix bugs
